views: moved the body contents to a template
This helloworld view achieves the following:

- dedicated reusable tag
- declarative template
- isolated behavior (localization)
- lazy loading

One limitation though, is the Content Security Policy kicking in when
the view's <script> tag code is present in the <polymer-element> tag.

For this reason, views scripts will be kept under `src/views` instead
of `src/js` to maintain locality of the templates and behaviors of
components.