views: moved the body contents to an external view
The view relies on xtag and is purely programmatic, the only downside
being that HTML templates are currently embedded inside JS code. OTOH,
xtag provides a powerful basis for view composition.

The use of RequireJS allows "lazy loading" of tag registrations (if we
ignore `r.js`) and dependency management between components.

The nice surprise is that l10n worked out-of-the-box!